+++ miscellaneoustool/data/src/main/kotlin/com/jinproject/data/di/BookDatabaseModule.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import androidx.room.Room
 import com.jinproject.core.util.doOnLocaleLanguage
 import com.jinproject.data.datasource.cache.database.BookDatabase
+import com.jinproject.data.datasource.cache.database.BookMigration
 import com.jinproject.data.datasource.cache.database.dao.CollectionDao
 import com.jinproject.data.datasource.cache.database.dao.DropListDao
 import com.jinproject.data.datasource.cache.database.dao.SimulatorDao
@@ -48,16 +49,20 @@ object BookDatabaseModule {
             onElse = "database/db_twom_3_eng.db"
         )
         val migration1to2 = context.doOnLocaleLanguage(
-            onKo = BookDatabase.MIGRATION_1_2_KOR,
-            onElse = BookDatabase.MIGRATION_1_2_ELSE
+            onKo = BookMigration.MIGRATION_1_2_KOR,
+            onElse = BookMigration.MIGRATION_1_2_ELSE
         )
         val migration2to3 = context.doOnLocaleLanguage(
-            onKo = BookDatabase.MIGRATION_2_3_KOR,
-            onElse = BookDatabase.MIGRATION_2_3_ELSE,
+            onKo = BookMigration.MIGRATION_2_3_KOR,
+            onElse = BookMigration.MIGRATION_2_3_ELSE,
+        )
+        val migration3to4 = context.doOnLocaleLanguage(
+            onKo = BookMigration.MIGRATION_3_4_KOR,
+            onElse = BookMigration.MIGRATION_3_4_ELSE,
         )
         return Room.databaseBuilder(context, BookDatabase::class.java,"BookDatabase")
             .createFromAsset(assetName)
-            .addMigrations(migration1to2, migration2to3)
+            .addMigrations(migration1to2, migration2to3, migration3to4)
             .build()
     }
 }
